+++ src/commands.cpp
@@ -143,8 +143,9 @@ void Commands::check(const std::string& path) {
   }
   Evaluator evaluator(settings);
   auto terms = seq.getTerms(OeisSequence::FULL_SEQ_LENGTH);
-  auto result =
-      evaluator.check(program, terms, OeisSequence::DEFAULT_SEQ_LENGTH, seq.id);
+  // don't use incremental evaluation in maintenance to detect bugs
+  auto result = evaluator.check(
+      program, terms, OeisSequence::DEFAULT_SEQ_LENGTH, seq.id, false);
   switch (result.first) {
     case status_t::OK:
       std::cout << "ok" << std::endl;