+++ src/main.cpp
@@ -140,7 +140,7 @@ int dispatch(Settings settings, const std::vector<std::string>& args) {
     Log::get().error("Option -s only allowed in evaluate command", true);
   }
   if (settings.print_as_b_file && cmd != "evaluate" && cmd != "eval" &&
-      cmd != "check" && cmd != "maintain") {
+      cmd != "check") {
     Log::get().error("Option -b not allowed for this command", true);
   }
   if (settings.parallel_mining && cmd != "mine") {